Refactor census3 queries
Following project structure see https://github.com/aragon/app/pull/1240#discussion_r1502407564
