+++ src/services/vocdoni-census3/census3-service.api.ts
@@ -1,0 +1,14 @@
+import {VocdoniCensus3Client, VocdoniSDKClient} from '@vocdoni/sdk';
+
+export interface IFetchVotingPowerByCensusId {
+  vocdoniClient: VocdoniSDKClient;
+  holderAddress: string;
+  censusId: string;
+}
+
+export interface IFetchVotingPowerByTokenAddress {
+  census3Client: VocdoniCensus3Client;
+  tokenId: string;
+  chainId: number;
+  holderId: string;
+}
