+++ src/services/vocdoni-census3/census3-service.ts
@@ -1,14 +1,17 @@
-import {VocdoniCensus3Client, VocdoniSDKClient} from '@vocdoni/sdk';
+import {
+  IFetchVotingPowerByCensusId,
+  IFetchVotingPowerByTokenAddress,
+} from './census3-service.api';
 
 /**
  * Returns the voting power for the specified address at specified vocdoni census id.
  * This is the way to get the voting power for vocdoni processes.
  */
-export async function getCensus3VotingPowerByCensusId(
-  vocdoniClient: VocdoniSDKClient,
-  holderAddress: string,
-  censusId: string
-) {
+export async function fetchVotingPowerByCensusId({
+  vocdoniClient,
+  holderAddress,
+  censusId,
+}: IFetchVotingPowerByCensusId) {
   return (await vocdoniClient.fetchProof(censusId, holderAddress)).weight;
 }
 
@@ -16,11 +19,11 @@ export async function getCensus3VotingPowerByCensusId(
  * Returns the voting power for the specified address using vocdoni census3 service.
  * It returns the last known holder balance for a specific token.
  */
-export async function getCensus3VotingPowerByTokenAddress(
-  census3Client: VocdoniCensus3Client,
-  tokenId: string,
-  chainId: number,
-  holderId: string
-) {
+export async function fetchVotingPowerByTokenAddress({
+  census3Client,
+  tokenId,
+  chainId,
+  holderId,
+}: IFetchVotingPowerByTokenAddress) {
   return await census3Client.tokenHolderBalance(tokenId, chainId, holderId);
 }
